pyproject.toml
[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"
source-exclude = ["bin"]
namespace = false

[project]
name = "django-revision"
version = "1.0.0"
description = "Track the git revision with every model instance saved"
readme = {file = "README.rst", content-type = "text/x-rst"}
dependencies = [
    'GitPython',
    "python-dateutil>=2.9.0.post0",
]
requires-python = '>=3.12'
keywords = ['Django', 'git', 'revision', 'clinicedc', 'edc']
authors = [{name="Erik van Widenfelt", email="ew2789@gmail.com"}]
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "GPL-3.0-or-later"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/erikvw/django-revision"
Documentation = "https://github.com/erikvw/django-revision/blob/README.rst"
Repository = "https://github.com/erikvw/django-revision.git"
Changelog = "https://github.com/erikvw/django-revision/blob/main/CHANGES"

[tool.black]
line-length = 88
target-version = ["py312"]
extend-exclude = '''^(.*\/)*\b(migrations|\.venv)\b($|\/.*$)'''

[tool.isort]
profile = "black"
py_version = "312"
skip = [".tox", ".eggs", "migrations", "_version.py", ".venv"]

[tool.coverage.run]
parallel = false
branch = true
source = ["src"]

[tool.coverage.paths]
source = ["drc"]

[tool.coverage.report]
show_missing = true
skip_covered = true
omit = ["requirements.txt"]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]

[tool.rstcheck]
ignore_directives = ["toctree", "code-block"]

[dependency-groups]
dev = [
    "django>=5.2",
    "ipython>=9.4.0",
    "python-dateutil>=2.9.0.post0",
    "edc-test-settings>=1.1.2",
]
docs = [
    "furo>=2025.7.19",
    "sphinx>=8.2.3",
    "sphinx-design>=0.6.1",
    "sphinx-inline-tabs>=2023.4.21",
]
lint = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort==5.13.2",
    "ruff>=0.12.5",
]

test = [
    "coverage>=7.10.3",
    "django>=5.2",
    "edc-test-settings>=1.1.2",
    "python-dateutil>=2.9.0.post0",
    "tox-uv>=1.28.0",
]

[tool.uv.sources]
django-revision = { workspace = true }
edc-test-settings = { path = "../edc-test-settings", editable = true }
edc-test-utils = { path = "../edc-test-utils", editable = true }
